requirements.txt
Flask
requests
google-generativeai
random 
python-dotenv
